Refactored the time spent calculation
